// js/stat-evaluation-player/src/generated/MovementEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type MovementEvent = { time: number, frame: number, player: RemoteIdTs, is_team_0: boolean, dt: number, speed: number, distance: number, speed_band: string, height_band: string, };